requirements.txt
numpy
Pillow
cryptography
tkinterdnd2
